+++ src/app.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import * as Joi from 'joi';
 import { AdminModule } from './admin/admin.module';
 import { AuthModule } from './auth/auth.module';
+import supabaseConfig from './environment/supabase.config';
+import webConfig from './environment/web.config';
 import { PrismaModule } from './prisma/prisma.module';
 import { ProfileModule } from './profile/profile.module';
-import supabaseConfig from './environment/supabase.config';
-import * as Joi from 'joi';
-import webConfig from './environment/web.config';
 
 @Module({
   imports: [
