+++ src/auth/auth.module.ts
@@ -1,14 +1,14 @@
-import { Global, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
+import { APP_GUARD } from '@nestjs/core';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { PassportModule } from '@nestjs/passport';
+import { PrismaClient } from '@prisma/client';
 import { createClient, SupabaseClient } from '@supabase/supabase-js';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { SupabaseStrategy } from './strategies/supabase.strategy';
-import { PassportModule } from '@nestjs/passport';
-import { APP_GUARD } from '@nestjs/core';
-import { SupabaseAuthGuard } from './guards/supabase.guard';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { SupabaseConfig } from '../environment';
-import { PrismaClient } from '@prisma/client';
+import { SupabaseConfig } from '../environment/supabase.config';
+import { SupabaseAuthGuard } from './guards';
+import { SupabaseStrategy } from './strategies';
 
 const options = {
   persistSession: false,
