+++ src/auth/guards/index.ts
@@ -1,0 +1,2 @@
+export * from './roles.guard';
+export * from './supabase.guard';
