--- src/environment/index.ts
@@ -1,2 +1,0 @@
-export * from './supabase.config';
-export * from './web.config';
