+++ src/profile/profile.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ProfileService } from './profile.service';
-import { ProfileController } from './profile.controller';
 import { PrismaClient } from '@prisma/client';
 import { AuthModule } from '../auth/auth.module';
+import { ProfileController } from './profile.controller';
+import { ProfileService } from './profile.service';
 
 @Module({
   controllers: [ProfileController],
